Mixed elasticity formualtion

+++ gsElThermoAssembler.hpp
@@ -15,10 +15,9 @@
 
 #include <gsElasticity/gsElThermoAssembler.h>
 
-
+#include <gsPde/gsPoissonPde.h>
 #include <gsElasticity/gsVisitorElThermo.h>
 #include <gsElasticity/gsVisitorElThermoBoundary.h>
-
 
 namespace gismo
 {
@@ -55,7 +54,7 @@ void gsElThermoAssembler<T>::findNonDirichletSides()
 {
     for (std::vector< patchSide >::iterator side = m_pde_ptr->domain().bBegin(); side != m_pde_ptr->domain().bEnd(); ++side)
     {
-        std::pair<int,boxSide> temp(side->patch,side->index());
+        std::pair<size_t,boxSide> temp(side->patch,side->index());
 
         typename gsBoundaryConditions<T>::const_iterator it = m_pde_ptr->bc().dirichletBegin();
         for ( ; it != m_pde_ptr->bc().dirichletEnd(); ++it )
